+++ src/components/card/card.vue
@@ -1,11 +1,17 @@
 <template>
-  <div :class="wrapperClasses">
-    <div v-if="props.showHeader && ($slots.header || props.title)" :class="headerClasses">
+  <div :class="cardClasses.baseClasses">
+    <div v-if="props.showHeader && ($slots.header || props.title)" :class="cardClasses.headerClasses">
       <div v-if="props.title" class="spr-flex spr-items-center">
-        <Icon v-if="props.headerIcon" :icon="props.headerIcon" class="spr-me-size-spacing-3xs spr-h-6 spr-w-6 spr-text-kangkong-600" />
+        <Icon
+          v-if="props.headerIcon"
+          :icon="props.headerIcon"
+          class="spr-me-size-spacing-3xs spr-h-6 spr-w-6 spr-text-kangkong-600"
+        />
         <div class="spr-mr-size-spacing-3xs">
           <div class="spr-text-mushroom-950 spr-font-size-300 spr-font-medium spr-leading-5">{{ props.title }}</div>
-          <div v-if="props.subtitle" class="spr-font-size-100 spr-leading-4 spr-text-mushroom-600">{{ props.subtitle }}</div>
+          <div v-if="props.subtitle" class="spr-font-size-100 spr-leading-4 spr-text-mushroom-600">
+            {{ props.subtitle }}
+          </div>
         </div>
       </div>
       <template v-if="$slots.header">
@@ -13,11 +19,13 @@
       </template>
     </div>
 
-    <div v-if="$slots.content" :class="contentPaddingSizeClasses">
+    <div v-if="$slots.content" :class="cardClasses.contentPaddingSizeClasses">
       <slot name="content" />
     </div>
+
     <slot />
-    <div v-if="$slots.footer && $props.showFooter" :class="footerClasses">
+
+    <div v-if="$slots.footer && $props.showFooter" :class="cardClasses.footerClasses">
       <slot name="footer" />
     </div>
   </div>
@@ -31,5 +39,5 @@ import { useCard } from './use-card';
 const props = defineProps(cardPropTypes);
 const slots = useSlots();
 
-const { wrapperClasses, headerClasses, footerClasses, contentPaddingSizeClasses } = useCard(props, slots);
+const { cardClasses } = useCard(props, slots);
 </script>