+++ src/components/input/input.ts
@@ -4,7 +4,7 @@ export const definePropType = <T>(val: unknown): PropType<T> => val as PropType<
 
 const TRAILING_SIZES = ['xs', 'sm', 'md'] as const;
 
-export const INPUT_TYPES = ['default','search', 'username', 'email', 'password', 'url'] as const;
+export const INPUT_TYPES = ['default', 'search', 'number','username', 'email', 'password', 'url'] as const;
 
 export const TYPE_HAS_TRAILING_ICONS = ['search', 'password', 'url'] as const;
 
@@ -26,7 +26,7 @@ export const inputPropTypes = {
   type: {
     type: String,
     validator: (value: (typeof INPUT_TYPES)[number]) => INPUT_TYPES.includes(value),
-    default: 'default',
+    default: 'text',
   },
   placeholder: {
     type: String,